backend/requirements.txt
# TradeIQ Backend - Design Document v3.0
# Optimized dependencies with version pins for stability

# Core Django Framework
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0

# Database
dj-database-url>=2.1,<3.0
psycopg2-binary>=2.9,<3.0

# WebSocket & Channels
channels>=4.0,<5.0
channels[daphne]>=4.0,<5.0
daphne>=4.0,<5.0

# AI/LLM
openai>=1.0.0,<2.0.0  # For DeepSeek API (OpenAI-compatible)

# External APIs
atproto>=0.0.18,<1.0.0  # Bluesky AT Protocol
requests>=2.31.0,<3.0.0  # For NewsAPI and external APIs

# Utilities
python-dotenv>=1.0,<2.0

# Authentication
PyJWT>=2.8,<3.0

# Optional: Development tools (uncomment if needed)
# ipython>=8.0,<9.0
# django-debug-toolbar>=4.0,<5.0
